% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gamtest.R
\name{print.gamtest}
\alias{print.gamtest}
\title{Print a gamtest Object}
\usage{
\method{print}{gamtest}(x, digits = 4, ...)
}
\arguments{
\item{x}{A gamtest object.}

\item{digits}{Number of digits for test statistic and p-value.}

\item{...}{Other generic options.}
}
\description{
This function print the semiparametric estimation of nonlinear curves and surface.
}
